fuse/fsdev: Fixed using setxattr flags and extended flags
Mount with xattrs enabled on fsdevs works now as well as setfacl and
getfacl.

What is done:
- Hanlding zero size (xattrs value request) in do_getxattr() so that it does
  not overflow and does not print error for this valid case of use.
- Support for compat size of the fuse_setxattr_in and crrectly getting
  setxattr name and value.
- Converting setxattr flags fuse->fsdev->posix. Extending flags to
  uint64_t and SPDK_FSDEV_SETXATTR_ACL_KILL_SGID to it
- Negotiating FUSE_SETXATTR_EXT and FUSE_POSIX_ACL mount flags.
- Using extended setxattr flags in lo_setxattr() to clear SGID on
  setting "system.posix_acl_access" in lo_setxattr() if dispatcher
  sends SPDK_FSDEV_SETXATTR_ACL_KILL_SGID in the flags.
- Fixed file open in AIO xattr-related functions:
  * Handling directories (required for acl).
  * Opening directories must be with O_RDONLY, the "readonly" flags is
    not related to xatts and works fine with fsetxattr().

+++ include/spdk/fsdev_module.h
@@ -361,7 +361,7 @@ struct spdk_fsdev_io {
 			char *name;
 			char *value;
 			size_t size;
-			uint32_t flags;
+			uint64_t flags;
 		} setxattr;
 		struct {
 			struct spdk_fsdev_file_object *fobject;
